Clean up serviceClient.ts by moving code
