+++ sdk/core/core-client/src/index.ts
@@ -3,12 +3,8 @@
 
 export { createSerializer, MapperTypeNames } from "./serializer";
 export { createSpanFunction } from "./createSpan";
-export {
-  ServiceClient,
-  ServiceClientOptions,
-  createClientPipeline,
-  ClientPipelineOptions
-} from "./serviceClient";
+export { ServiceClient, ServiceClientOptions } from "./serviceClient";
+export { createClientPipeline, ClientPipelineOptions } from "./pipeline";
 export {
   OperationSpec,
   OperationArguments,
