+++ sdk/core/core-client/src/serviceClient.ts
@@ -7,28 +7,16 @@ import {
   PipelineRequest,
   PipelineResponse,
   Pipeline,
-  createPipelineRequest,
-  createPipelineFromOptions,
-  bearerTokenAuthenticationPolicy,
-  InternalPipelineOptions
+  createPipelineRequest
 } from "@azure/core-https";
-import {
-  OperationArguments,
-  OperationSpec,
-  OperationRequest,
-  OperationResponseMap,
-  FullOperationResponse,
-  CompositeMapper,
-  XmlOptions
-} from "./interfaces";
+import { OperationArguments, OperationSpec, OperationRequest, XmlOptions } from "./interfaces";
 import { getStreamingResponseStatusCodes } from "./interfaceHelpers";
 import { getRequestUrl } from "./urlHelpers";
-import { isPrimitiveType } from "./utils";
-import { deserializationPolicy, DeserializationPolicyOptions } from "./deserializationPolicy";
+import { flattenResponse } from "./utils";
 import { URL } from "./url";
-import { serializationPolicy, serializationPolicyOptions } from "./serializationPolicy";
 import { getCachedDefaultHttpsClient } from "./httpClientCache";
 import { getOperationRequestInfo } from "./operationHelpers";
+import { createClientPipeline } from "./pipeline";
 
 /**
  * Options to be provided while creating the client.
@@ -242,116 +230,6 @@ function createDefaultPipeline(
       stringifyXML: options.stringifyXML
     }
   });
-}
-
-/**
- * Options for creating a Pipeline to use with ServiceClient.
- * Mostly for customizing the auth policy (if using token auth) or
- * the deserialization options when using XML.
- */
-export interface ClientPipelineOptions extends InternalPipelineOptions {
-  /**
-   * Options to customize bearerTokenAuthenticationPolicy.
-   */
-  credentialOptions?: { credentialScopes: string | string[]; credential: TokenCredential };
-  /**
-   * Options to customize deserializationPolicy.
-   */
-  deserializationOptions?: DeserializationPolicyOptions;
-  /**
-   * Options to customize serializationPolicy.
-   */
-  serializationOptions?: serializationPolicyOptions;
-}
-
-/**
- * Creates a new Pipeline for use with a Service Client.
- * Adds in deserializationPolicy by default.
- * Also adds in bearerTokenAuthenticationPolicy if passed a TokenCredential.
- * @param options - Options to customize the created pipeline.
- */
-export function createClientPipeline(options: ClientPipelineOptions = {}): Pipeline {
-  const pipeline = createPipelineFromOptions(options ?? {});
-  if (options.credentialOptions) {
-    pipeline.addPolicy(
-      bearerTokenAuthenticationPolicy({
-        credential: options.credentialOptions.credential,
-        scopes: options.credentialOptions.credentialScopes
-      })
-    );
-  }
-
-  pipeline.addPolicy(serializationPolicy(options.serializationOptions), { phase: "Serialize" });
-  pipeline.addPolicy(deserializationPolicy(options.deserializationOptions), {
-    phase: "Deserialize"
-  });
-
-  return pipeline;
-}
-
-function flattenResponse(
-  fullResponse: FullOperationResponse,
-  responseSpec: OperationResponseMap | undefined
-): unknown {
-  const parsedHeaders = fullResponse.parsedHeaders;
-  const bodyMapper = responseSpec && responseSpec.bodyMapper;
-
-  if (bodyMapper) {
-    const typeName = bodyMapper.type.name;
-    if (typeName === "Stream") {
-      return {
-        ...parsedHeaders,
-        blobBody: fullResponse.blobBody,
-        readableStreamBody: fullResponse.readableStreamBody
-      };
-    }
-
-    const modelProperties =
-      (typeName === "Composite" && (bodyMapper as CompositeMapper).type.modelProperties) || {};
-    const isPageableResponse = Object.keys(modelProperties).some(
-      (k) => modelProperties[k].serializedName === ""
-    );
-    if (typeName === "Sequence" || isPageableResponse) {
-      const arrayResponse: { [key: string]: unknown } =
-        fullResponse.parsedBody ?? (([] as unknown) as { [key: string]: unknown });
-
-      for (const key of Object.keys(modelProperties)) {
-        if (modelProperties[key].serializedName) {
-          arrayResponse[key] = fullResponse.parsedBody?.[key];
-        }
-      }
-
-      if (parsedHeaders) {
-        for (const key of Object.keys(parsedHeaders)) {
-          arrayResponse[key] = parsedHeaders[key];
-        }
-      }
-      return arrayResponse;
-    }
-
-    if (typeName === "Composite" || typeName === "Dictionary") {
-      return {
-        ...parsedHeaders,
-        ...fullResponse.parsedBody
-      };
-    }
-  }
-
-  if (
-    bodyMapper ||
-    fullResponse.request.method === "HEAD" ||
-    isPrimitiveType(fullResponse.parsedBody)
-  ) {
-    return {
-      ...parsedHeaders,
-      body: fullResponse.parsedBody
-    };
-  }
-
-  return {
-    ...parsedHeaders,
-    ...fullResponse.parsedBody
-  };
 }
 
 function getCredentialScopes(options: ServiceClientOptions): string | string[] | undefined {
