+++ sdk/core/core-client/src/utils.ts
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+import { CompositeMapper, FullOperationResponse, OperationResponseMap } from "./interfaces";
+
 /**
  * Returns true if the given value is a basic/primitive type
  * (string, number, boolean, null, undefined).
@@ -33,4 +35,75 @@ const validUuidRegex = /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F
  */
 export function isValidUuid(uuid: string): boolean {
   return validUuidRegex.test(uuid);
+}
+
+/**
+ * Take a `FullOperationResponse` and turn it into a flat
+ * response object to hand back to the consumer.
+ * @param fullResponse The processed response from the operation request
+ * @param responseSpec The response map from the OperationSpec
+ */
+export function flattenResponse(
+  fullResponse: FullOperationResponse,
+  responseSpec: OperationResponseMap | undefined
+): unknown {
+  const parsedHeaders = fullResponse.parsedHeaders;
+  const bodyMapper = responseSpec && responseSpec.bodyMapper;
+
+  if (bodyMapper) {
+    const typeName = bodyMapper.type.name;
+    if (typeName === "Stream") {
+      return {
+        ...parsedHeaders,
+        blobBody: fullResponse.blobBody,
+        readableStreamBody: fullResponse.readableStreamBody
+      };
+    }
+
+    const modelProperties =
+      (typeName === "Composite" && (bodyMapper as CompositeMapper).type.modelProperties) || {};
+    const isPageableResponse = Object.keys(modelProperties).some(
+      (k) => modelProperties[k].serializedName === ""
+    );
+    if (typeName === "Sequence" || isPageableResponse) {
+      const arrayResponse: { [key: string]: unknown } =
+        fullResponse.parsedBody ?? (([] as unknown) as { [key: string]: unknown });
+
+      for (const key of Object.keys(modelProperties)) {
+        if (modelProperties[key].serializedName) {
+          arrayResponse[key] = fullResponse.parsedBody?.[key];
+        }
+      }
+
+      if (parsedHeaders) {
+        for (const key of Object.keys(parsedHeaders)) {
+          arrayResponse[key] = parsedHeaders[key];
+        }
+      }
+      return arrayResponse;
+    }
+
+    if (typeName === "Composite" || typeName === "Dictionary") {
+      return {
+        ...parsedHeaders,
+        ...fullResponse.parsedBody
+      };
+    }
+  }
+
+  if (
+    bodyMapper ||
+    fullResponse.request.method === "HEAD" ||
+    isPrimitiveType(fullResponse.parsedBody)
+  ) {
+    return {
+      ...parsedHeaders,
+      body: fullResponse.parsedBody
+    };
+  }
+
+  return {
+    ...parsedHeaders,
+    ...fullResponse.parsedBody
+  };
 }
